refactor IJDDItem
