+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/hierarchy/FileTreeHierarchyGenerator.kt
@@ -11,7 +11,7 @@ import org.plan.research.minimization.plugin.errors.HierarchyBuildError
 import org.plan.research.minimization.plugin.errors.NoExceptionFound
 import org.plan.research.minimization.plugin.errors.NoRootFound
 import org.plan.research.minimization.plugin.execution.SameExceptionPropertyTester
-import org.plan.research.minimization.plugin.model.IJDDItem.VirtualFileDDItem
+import org.plan.research.minimization.plugin.model.VirtualFileDDItem
 import org.plan.research.minimization.plugin.model.ProjectHierarchyProducer
 import org.plan.research.minimization.plugin.services.CompilationPropertyCheckerService
